+++ packages/gamut-styles/src/index.ts
@@ -3,6 +3,8 @@ import '@emotion/react';
 import { theme } from './theme';
 
 export * from './GamutProvider';
+export * from './ColorMode';
+export * from './Background';
 
 export * from './variance';
 export * from './cache';